Create functions for writing simulation data to file
Created functions for writing simulation data to file.
The writing is similar to how the experimental data is, though
currently not an exact match, and in general just a quick setup.
It should still allow easy creation of multiple simulated data and
writing the data along with the metadata to file.