Add function to get somewhat randomised input parameters
As per the title. Basically looked at the real data parameters
and added some variation based on those to the creation of the
input parameters for the simulation.

Idea at this point is to have some way to generate somewhat
randomised data to see whether a neural network is able to learn
well enough. Obviously some better randomisation should be done
ultimately. For now, it's more important to be able to generate
enough data that is somewhat randomised but also looks like it could
have been generated by a similar process as with the real data.