Add reading of simulated data and its metadata
It should be possible to also read the real data
with the read function, but the metadata for the
real data cannot currently be read with the metadata
function. It's perhaps best to just transform the metadata
for the real data into the same JSON format the simdata is
currently in, after which it would also be possible to
read the real data's metadata in with the function.