Add simple test for MLPRegressor
At this point, the idea was to just test
the overall functionality of everything.
Data seems to read in okay, MLPRegressor
seems to run, though does not complete due
to too few iterations.

Need to still do some stuff with the data
to have it in a better form for the regressor.