Add preprocessing for input and output data
Scaling the data seems to be helping the prediction somewhat.
There might be some better way to do the preprocessing, perhaps
in sklearn, a way that would allow easy transforming both ways.
For now, however, the added functions suffice, though the reverse
transformations have yet to be added.

It might be good to add something to make it clear that the
intensities sum up to one. Would need the background as well
for this, though.

Testing against the real data is also yet to be done, which is
quite crucial as well.