Change data reading to accomodate real data
There was something odd going on with the real data that
it did not want to be read in properly, but randomly used column
values as index. In any case, needed to ignore the header line, instead
explicitly passing in the headers (less convenient, I'd say) and also
disabled getting indices from the data itself, which is of course fine.