Add MLP in PyTorch
This was done to have more control over the process, which should allow
to better refine the model and to actually see what is happening
"inside". In comparison, the sklearn MLPRegressor does not allow
for much information about the fitting process, and the documentation is
lacking on how exactly the regressor works.