Improve the way many simulations are written to file
The last written to index is now returned and can be given as argument,
so that the next one doesn't always have to be searched for. Not sure
how much of a problem this was anyway. The writing's still not that
fast, but doesn't ultimately take that long.

Keeping the metadata file open between writes would probably be helpful
as well, but currently too much of a hassle to implement.