Change model saving to work correctly and better
Previously the model would've been saved in the final state as the
variable was a reference. Using deepcopy, it should now be a
detached state_dict of the "best" performing model state. In addition,
the model layer sizes are also saved with the state_dict, as well as
the normalisation.