Separate previous model training function into smaller parts
The earlier model_training() from pytorchMLP_test was quite a behemoth.
The idea is to cut it into more digestible pieces, and obviously
generally make it all a little more modular by having the different
functionalities in their own functions.

Not quite done with the whole process, but it's a start.