Patch the previous commit to have rng still speficiable for functions
Realised immediately that to actually have a seedable rng outside
the module, an rng argument would still be needed in the function arguments.
There's probably some workaround for this, but the numpy rng generators
don't seem to allow seeding after the fact, for one, meaning it's
probably quite difficult to seed it once and then have it generate
with that seed continuously going forward. Also changing things inside
a module from another module seems to be a shaky thing to do.

Well, difficulties or not, it seems the simplest way to do it is
by allowing rngs as argument. The module itself can still use a
module-wide rng, so no rng has to be passed around when adding
functions there. Though at the same time, it in fact now needs
even more code to support it all, most importantly requiring that the
rng arguments are written in anyway, and checked for. So probably
kind of a dumb change.

A simple workaround would be to make the module into a class, I guess.
Then the rng could be held in the class and changed at any time,
but would not need to be passed around constantly.