Change the way simulation data is written to file
This will also change the formatting of the data files, and remove
the use of the metadata file. The new formatting has the components
in the dictionary (or json) format on the first line, and the rest
of the file is the time and counts; each simulated data is now
self-contained. Still need to change the read functions to reflect
this change.

A preliminary test seems to suggest that currently, each write
takes around 0.06 seconds, and this does not appear to be
immediately dependent on the number of files already written in the
specified folder. This is perhaps not surprising, but is a change
in comparison to the previous case, where seeking the evergrowing
metadata file each write made the simulation writing slower the more
data there was in the folder, presumably.