Add function for reading new format of simulation data
In relation to the previous commit. Now both the counts and the
components are read and returned at the same time, as they're both in
the same file. Seems like it might be a little faster to read the data
as well, but that remains to be seen.

Still need to update things elsewhere (train.py, evaluate.py) to
reflect these changes.