Update training to use new simulation formatting, fix return from data read
The training main function now reads the simdata with the new read function.

Had to change the data read function to return the counts as a list
of arrays, as the processing function currently expects it in this
format. Decided it would be best for now to not change the processing
so older style data reads elsewhere don't stop working. More specifically,
the experimental data is still in the old format, kind of, and would
be read in with the old functionalities.

Ideally should format the experimental data to the new format, which
shouldn't take much, and then possibly fix up the other parts. Probably
best to make a PyTorch dataloader outright.