Update README, add instructions to files, polish functions
Updated README to have some better instructions, and added
instructions for running the simulation, training and evaluation.

Main change on the function side was in read_data.py, where the
data file fetching was changed to be hopefully a little easier to use.
Also moved some of the variables from the main functions to be arguments
to the main functions for easier usage.