Add use of CUDA capabilities and choosing dtype
It looks like using CUDA does help a little: running with CPU gave
roughly 13 seconds when training, while with CUDA it's around 9 seconds.
This was with only a hundred epochs.

The choice of dtype was also added as often float16 has been necessary,
but it did not seem to be here. Also, float16 seemed to cause some
problems, presumably due to the limited precision, such that the results
were all NaN, so it's best to stick to float32 if possible
until the problem is found.