Add input processing parameters to saved train data
Previously no saving of these parameters was done, which meant
that when evaluating, the user would just have to know which
parameters they had used (although this may have been possible to
infer from the logs). Now this should be fixed, such that when
saving the model in training, the processing input parameters used in
training are also saved for later use.