Add generic class for neural nets, add CNN for first time
Not necessarily the most stable commit, slightly a checkpoint.

Currently, the Conv1 that adds a little to the base torch.nn.Conv1d could be
problematic: it needs to expand input to a fitting form, which must
be transformed back. If it's not, I'm not sure what a linear layer,
for example, would do (ideally throw an exception, but it might just
happily churn through). However, as long as a flatten starting at
dimension 1 is done in between, it actually should not be a problem.
(the conv_compress should be changed to this, probably.)