Improve active plotting usability
Previously each axis could only have one bit of data plotted to
it, but it's actually easy enough to just access all the "artists"
of the axes, so it's now possible to plot any number of plots (supposedly)
in any given axis.

Overall, the whole thing could still use improvements, but again,
not of utmost importance right now.