Change simulation writing to use integers
Realised that as the writing used floats in the scientific form and
had the whole float shown, it makes up quite a lot of unnecessary
data, so that the file sizes are bloated. This helps to fix the issue.
However, the idea would be to ultimately save the data in the PyTorch
.pt binary format.