Add option for choosing which simulation input parameters to fit to
Also a lot of other changes that needed to be made as a result, plus
other stuff that should have been committed a while ago. Not great git
management, and really should be using dev branches, so I'd not have
to commit broken stuff to main. Also actually really ensuring that
everything relevant is committed.