Add methods for handling more complex model outputs
At the moment, these should not change that much how the whole thing runs.
The idea behind the changes is that for predicting mean and variance,
the handling of calculating the loss needs to be slightly different.
Basically, the "mean" is the actual prediction of the true values,
and that might be more of interest in some cases for comparisons,
but for calculating a loss like GaussianNLLLoss, both the mean
and variance is required, and cannot also be simply given to the
loss function the same way a prediction would be given to MSELoss.
As such, to maintain the interface, a prediction output from a model
should be given to the wrapper of the loss function as-is (such that
only the one variable from the model and the true value is given to the
wrapper, rather than a mean and variance and the true, for example). The
wrapper then handles the given values as necessary, for example separating
the mean and variance from the model prediction output, and handing
those values, along with the true values, to the GaussianNLLLoss.

A separate method can be overwritten to only output the means in the above
case, useful, for one, in evaluating the mean predictions.