Change PALS network class to support better subclassing
Had a long think as to how best allow adding the GaussianNLLLoss.
One idea would be to instead pass in a dictionary, which has tuples
of functions and indices as values. This would be iterated over,
with each function being applied to the output of the last layer
based on the indices matching the function. The output from this
would be put in another dictionary with keys corresponding to
the (function, indices) dictionary. This way, the class is as general
as possible. The downside is that firstly, evaluating the functions
might either require evaluating identity functions or checking for
a None, which would take unnecessary time. Furthermore, it's not
a very clear way to keep track of how a specific network has functioned,
if the only difference is a user defined dictionary given as input to
the __init__.

By allowing proper subclassing of the current class, the network is given an actual
name, avoiding the problem of the ephemeral user-defined dictionary,
though the class itself may change over time anyway. Still, the interface
is likely to be more clear. Additionally,
subclassing the class instead means that the subclass implementation
can be as efficient as possible, with no unnecessary ifs or function
evaluations that would arise in the former considered case. In turn,
doing all this does mean that there can be a perhaps unnecessary amount
of various slightly different network classes, the functionality of
which could be reduced to just one class, but such is the nature of
OOP, I suppose.