Simplify model loading and move it to separate file
The simplification is that now, the network's class is saved with
the weights and other training info, so it's easier to load up
a model. The big difference is that now it's easy enough to also continue
training from a saved state. A thing to add would be saving the whole model environment,
with the state of the optimiser and scheduler and such, which is probably
not too difficult, but I'll keep this commit smaller.