Modify train to allow loading the whole model or the network
Might also want to add the ability to separately load certain
states of the entire model. The main issue is that loading
in just the network means that the state of the optimiser is not loaded,
which is an issue with Adam, for example, as it has its moments. In turn,
loading the whole model means it's currently not possible to also
define a new learning rate, for example, on the fly. Of course,
this is a more general thing, and could be solved otherwise,
as it could be nice to be able to dynamically choose the learning
rate whenever.