Add classes for processing output
Realised that actually, it's probably better to subtract the minimum
and then divide the result by the result's maximum, rather
than just dividing by the maximum, as then the range of each output
value is actually [0,1], which may make training a little more stable,
and seems to do. In light of this, felt that it was best to make
a more rigorous way to do processing, both during training and
later during evaluation. Now, much less needs to be known by the
user about the processing, so it's more clear and easier to do,
especially as the processor class and it's state can be saved during training.