Change the way softplus is computed in the network
I had understood the working of the softplus slightly
wrong, thinking it was not exactly an approximation of the ReLU
and had to be multiplied by a constant. Upon further inspection,
I realised this was not the case, and no multiplication is necessary.

Still, this should not have had any meaningful effect on the training
of the networks.